move test to form_fields tests
